Move EnemyPokemon specific init Logic to its class
